fix: separate redshift workspace test from snowflake workspace test, because it cannot run on ABS backend
