+++ Tests/Reader/DownloadTablesWorkspaceRedshiftTest.php
@@ -1,0 +1,65 @@
+<?php
+
+namespace Keboola\InputMapping\Tests\Reader;
+
+use Keboola\InputMapping\Configuration\Table\Manifest\Adapter;
+use Keboola\InputMapping\Reader\Options\InputTableOptionsList;
+use Keboola\InputMapping\Reader\Reader;
+use Keboola\InputMapping\Reader\State\InputTableStateList;
+use Keboola\StorageApi\Client;
+use Psr\Log\Test\TestLogger;
+
+class DownloadTablesWorkspaceRedshiftTest extends DownloadTablesWorkspaceTestAbstract
+{
+    public function testTablesRedshiftBackend()
+    {
+        $logger = new TestLogger();
+        $reader = new Reader($this->client, $logger, $this->getWorkspaceProvider());
+        $configuration = new InputTableOptionsList([
+            [
+                'source' => 'in.c-input-mapping-test.test1',
+                'destination' => 'test1',
+                'changed_since' => '-2 days',
+            ],
+            [
+                'source' => 'in.c-input-mapping-test.test2',
+                'destination' => 'test2',
+            ]
+        ]);
+
+        $reader->downloadTables(
+            $configuration,
+            new InputTableStateList([]),
+            $this->temp->getTmpFolder() . DIRECTORY_SEPARATOR . 'download',
+            'workspace-redshift'
+        );
+
+        $adapter = new Adapter();
+
+        $manifest = $adapter->readFromFile($this->temp->getTmpFolder() . '/download/test1.manifest');
+        /* because of https://keboola.atlassian.net/browse/KBC-228 we have to create redshift bucket to
+            unload data from redshift workspace */
+        $this->client->dropBucket('out.c-input-mapping-test');
+        $this->client->createBucket('input-mapping-test', Client::STAGE_OUT, 'Docker Testsuite', 'redshift');
+
+        self::assertEquals('in.c-input-mapping-test.test1', $manifest['id']);
+        // test that the table exists in the workspace
+        $tableId = $this->client->createTableAsyncDirect(
+            'out.c-input-mapping-test',
+            ['dataWorkspaceId' => $this->workspaceId, 'dataTableName' => 'test1', 'name' => 'test1']
+        );
+        self::assertEquals('out.c-input-mapping-test.test1', $tableId);
+
+        $manifest = $adapter->readFromFile($this->temp->getTmpFolder() . '/download/test2.manifest');
+        self::assertEquals('in.c-input-mapping-test.test2', $manifest['id']);
+        $tableId = $this->client->createTableAsyncDirect(
+            'out.c-input-mapping-test',
+            ['dataWorkspaceId' => $this->workspaceId, 'dataTableName' => 'test2', 'name' => 'test2']
+        );
+        self::assertEquals('out.c-input-mapping-test.test2', $tableId);
+
+        self::assertTrue($logger->hasInfoThatContains('Table "in.c-input-mapping-test.test1" will be copied.'));
+        self::assertTrue($logger->hasInfoThatContains('Table "in.c-input-mapping-test.test2" will be copied.'));
+        self::assertTrue($logger->hasInfoThatContains('Processing 1 workspace exports.'));
+    }
+}
