+++ Tests/Reader/DownloadTablesWorkspaceSnowflakeTest.php
@@ -2,83 +2,16 @@
 
 namespace Keboola\InputMapping\Tests\Reader;
 
-use Keboola\Csv\CsvFile;
 use Keboola\InputMapping\Configuration\Table\Manifest\Adapter;
 use Keboola\InputMapping\Exception\InvalidInputException;
-use Keboola\InputMapping\Reader\NullWorkspaceProvider;
 use Keboola\InputMapping\Reader\Options\InputTableOptionsList;
 use Keboola\InputMapping\Reader\Reader;
 use Keboola\InputMapping\Reader\State\InputTableStateList;
-use Keboola\InputMapping\Reader\WorkspaceProviderInterface;
-use Keboola\StorageApi\Client;
 use Keboola\StorageApi\ClientException;
-use Keboola\StorageApi\Workspaces;
 use Psr\Log\Test\TestLogger;
 
-class DownloadTablesWorkspaceTest extends DownloadTablesTestAbstract
+class DownloadTablesWorkspaceSnowflakeTest extends DownloadTablesWorkspaceTestAbstract
 {
-    private $workspaceId;
-
-    public function setUp()
-    {
-        parent::setUp();
-        try {
-            $this->client->dropBucket('in.c-input-mapping-test', ['force' => true]);
-        } catch (ClientException $e) {
-            if ($e->getCode() !== 404) {
-                throw $e;
-            }
-        }
-        try {
-            $this->client->dropBucket('out.c-input-mapping-test', ['force' => true]);
-        } catch (ClientException $e) {
-            if ($e->getCode() !== 404) {
-                throw $e;
-            }
-        }
-        $this->client->createBucket('input-mapping-test', Client::STAGE_IN, 'Docker Testsuite');
-        $this->client->createBucket('input-mapping-test', Client::STAGE_OUT, 'Docker Testsuite');
-
-        // Create table
-        $csv = new CsvFile($this->temp->getTmpFolder() . DIRECTORY_SEPARATOR . 'upload.csv');
-        $csv->writeRow(['Id', 'Name', 'foo', 'bar']);
-        $csv->writeRow(['id1', 'name1', 'foo1', 'bar1']);
-        $csv->writeRow(['id2', 'name2', 'foo2', 'bar2']);
-        $csv->writeRow(['id3', 'name3', 'foo3', 'bar3']);
-        $this->client->createTableAsync('in.c-input-mapping-test', 'test1', $csv);
-        $this->client->createTableAsync('in.c-input-mapping-test', 'test2', $csv);
-        $this->client->createTableAsync('in.c-input-mapping-test', 'test3', $csv);
-    }
-
-    public function tearDown()
-    {
-        if ($this->workspaceId) {
-            $workspaces = new Workspaces($this->client);
-            $workspaces->deleteWorkspace($this->workspaceId);
-            $this->workspaceId = null;
-        }
-        parent::tearDown();
-    }
-
-    private function getWorkspaceProvider()
-    {
-        $mock = self::getMockBuilder(NullWorkspaceProvider::class)
-            ->setMethods(['getWorkspaceId'])
-            ->getMock();
-        $mock->method('getWorkspaceId')->willReturnCallback(
-            function ($type) {
-                if (!$this->workspaceId) {
-                    $workspaces = new Workspaces($this->client);
-                    $workspace = $workspaces->createWorkspace(['backend' => $type]);
-                    $this->workspaceId = $workspace['id'];
-                }
-                return $this->workspaceId;
-            }
-        );
-        /** @var WorkspaceProviderInterface $mock */
-        return $mock;
-    }
-
     public function testTablesSnowflakeBackend()
     {
         $logger = new TestLogger();
@@ -163,58 +96,6 @@ class DownloadTablesWorkspaceTest extends DownloadTablesTestAbstract
         self::assertEquals(2, count($params['input']));
         self::assertEquals('test1', $params['input'][0]['destination']);
         self::assertEquals('test3', $params['input'][1]['destination']);
-    }
-
-    public function testTablesRedshiftBackend()
-    {
-        $logger = new TestLogger();
-        $reader = new Reader($this->client, $logger, $this->getWorkspaceProvider());
-        $configuration = new InputTableOptionsList([
-            [
-                'source' => 'in.c-input-mapping-test.test1',
-                'destination' => 'test1',
-                'changed_since' => '-2 days',
-            ],
-            [
-                'source' => 'in.c-input-mapping-test.test2',
-                'destination' => 'test2',
-            ]
-        ]);
-
-        $reader->downloadTables(
-            $configuration,
-            new InputTableStateList([]),
-            $this->temp->getTmpFolder() . DIRECTORY_SEPARATOR . 'download',
-            'workspace-redshift'
-        );
-
-        $adapter = new Adapter();
-
-        $manifest = $adapter->readFromFile($this->temp->getTmpFolder() . '/download/test1.manifest');
-        /* because of https://keboola.atlassian.net/browse/KBC-228 we have to create redshift bucket to
-            unload data from redshift workspace */
-        $this->client->dropBucket('out.c-input-mapping-test');
-        $this->client->createBucket('input-mapping-test', Client::STAGE_OUT, 'Docker Testsuite', 'redshift');
-
-        self::assertEquals('in.c-input-mapping-test.test1', $manifest['id']);
-        // test that the table exists in the workspace
-        $tableId = $this->client->createTableAsyncDirect(
-            'out.c-input-mapping-test',
-            ['dataWorkspaceId' => $this->workspaceId, 'dataTableName' => 'test1', 'name' => 'test1']
-        );
-        self::assertEquals('out.c-input-mapping-test.test1', $tableId);
-
-        $manifest = $adapter->readFromFile($this->temp->getTmpFolder() . '/download/test2.manifest');
-        self::assertEquals('in.c-input-mapping-test.test2', $manifest['id']);
-        $tableId = $this->client->createTableAsyncDirect(
-            'out.c-input-mapping-test',
-            ['dataWorkspaceId' => $this->workspaceId, 'dataTableName' => 'test2', 'name' => 'test2']
-        );
-        self::assertEquals('out.c-input-mapping-test.test2', $tableId);
-
-        self::assertTrue($logger->hasInfoThatContains('Table "in.c-input-mapping-test.test1" will be copied.'));
-        self::assertTrue($logger->hasInfoThatContains('Table "in.c-input-mapping-test.test2" will be copied.'));
-        self::assertTrue($logger->hasInfoThatContains('Processing 1 workspace exports.'));
     }
 
     public function testTablesInvalidMapping()
