+++ phpunit.xml.dist
@@ -16,10 +16,12 @@
             <exclude>Tests/Reader/DownloadFilesRedshiftTest.php</exclude>
             <exclude>Tests/Reader/DownloadTablesRedshiftTest.php</exclude>
             <exclude>Tests/Reader/DownloadTablesS3DefaultTest.php</exclude>
+            <exclude>Tests/Reader/DownloadTablesWorkspaceRedshiftTest.php</exclude>
         </testsuite>
         <testsuite name="Redshift">
             <file>Tests/Reader/DownloadFilesRedshiftTest.php</file>
             <file>Tests/Reader/DownloadTablesRedshiftTest.php</file>
+            <file>Tests/Reader/DownloadTablesWorkspaceRedshiftTest.php</file>
         </testsuite>
         <testsuite name="S3">
             <file>Tests/Reader/DownloadTablesS3DefaultTest.php</file>
